improve algorithm
